Move Exemplar and AttachmentValue to metrics.data.
