+++ api/src/main/java/io/opencensus/metrics/export/Distribution.java
@@ -17,10 +17,10 @@
 package io.opencensus.metrics.export;
 
 import com.google.auto.value.AutoValue;
-import io.opencensus.common.Exemplar;
 import io.opencensus.common.ExperimentalApi;
 import io.opencensus.common.Function;
 import io.opencensus.internal.Utils;
+import io.opencensus.metrics.data.Exemplar;
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
