+++ api/src/test/java/io/opencensus/metrics/data/ExemplarTest.java
@@ -14,11 +14,12 @@
  * limitations under the License.
  */
 
-package io.opencensus.common;
+package io.opencensus.metrics.data;
 
 import static com.google.common.truth.Truth.assertThat;
 
-import io.opencensus.common.AttachmentValue.AttachmentValueString;
+import io.opencensus.common.Timestamp;
+import io.opencensus.metrics.data.AttachmentValue.AttachmentValueString;
 import java.util.Collections;
 import java.util.Map;
 import org.junit.Rule;
@@ -27,7 +28,7 @@ import org.junit.rules.ExpectedException;
 import org.junit.runner.RunWith;
 import org.junit.runners.JUnit4;
 
-/** Unit tests for {@link Exemplar}. */
+/** Unit tests for {@link io.opencensus.metrics.data.Exemplar}. */
 @RunWith(JUnit4.class)
 public class ExemplarTest {
