+++ buildscripts/import-control.xml
@@ -90,6 +90,7 @@ General guidelines on imports:
       <allow pkg="io.opencensus.trace"/>
     </subpackage>
     <subpackage name="exemplar.util">
+      <allow pkg="io.opencensus.metrics.data"/>
       <allow pkg="io.opencensus.stats"/>
       <allow pkg="io.opencensus.trace"/>
     </subpackage>
