+++ contrib/exemplar_util/src/main/java/io/opencensus/contrib/exemplar/util/ExemplarUtils.java
@@ -16,7 +16,7 @@
 
 package io.opencensus.contrib.exemplar.util;
 
-import io.opencensus.common.Exemplar;
+import io.opencensus.metrics.data.Exemplar;
 import io.opencensus.stats.MeasureMap;
 import io.opencensus.trace.SpanContext;
 import javax.annotation.Nullable;
