+++ exporters/metrics/ocagent/src/main/java/io/opencensus/exporter/metrics/ocagent/MetricsProtoUtils.java
@@ -18,11 +18,11 @@ package io.opencensus.exporter.metrics.ocagent;
 
 import com.google.protobuf.DoubleValue;
 import com.google.protobuf.Int64Value;
-import io.opencensus.common.AttachmentValue;
-import io.opencensus.common.Exemplar;
 import io.opencensus.common.Function;
 import io.opencensus.common.Functions;
 import io.opencensus.common.Timestamp;
+import io.opencensus.metrics.data.AttachmentValue;
+import io.opencensus.metrics.data.Exemplar;
 import io.opencensus.metrics.export.Distribution;
 import io.opencensus.metrics.export.Summary;
 import io.opencensus.proto.metrics.v1.DistributionValue;
