+++ impl_core/src/main/java/io/opencensus/implcore/stats/MeasureToViewMap.java
@@ -20,10 +20,10 @@ import com.google.common.collect.HashMultimap;
 import com.google.common.collect.Maps;
 import com.google.common.collect.Multimap;
 import com.google.common.collect.Sets;
-import io.opencensus.common.AttachmentValue;
 import io.opencensus.common.Clock;
 import io.opencensus.common.Timestamp;
 import io.opencensus.implcore.internal.CurrentState.State;
+import io.opencensus.metrics.data.AttachmentValue;
 import io.opencensus.metrics.export.Metric;
 import io.opencensus.stats.Measure;
 import io.opencensus.stats.Measurement;
